requirements.txt
python-telegram-bot==20.3
Pillow==9.5.0
gTTS==2.3.1
pytesseract==0.3.10
comtypes==1.1.10
PyMuPDF==1.22.3

